requirements.txt
flask
flask-socketio
mysql-connector-python
requests
python-socketio